+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Unary minus for `SpectralFunction`.
+- `models::bethe()`, density of states of a Bethe lattice with a finite coordination
+  number.
 
 ## [0.1.0] - 2026-08-28